third_party/blink/web_tests/platform/android/external/wpt/webnn/idlharness.https.any.worker-expected.txt

This is a testharness.js-based test.
Found 324 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] idl_test setup
  promise_test: Unhandled rejection with value: object "ReferenceError: MLGraphBuilder is not defined"
[FAIL] NavigatorML must be primary interface of navigator
  assert_own_property: self does not have own property "NavigatorML" expected property "NavigatorML" missing
[FAIL] Stringification of navigator
  assert_class_string: class string of navigator expected "[object NavigatorML]" but got "[object WorkerNavigator]"
[FAIL] NavigatorML interface: navigator must not have property "ml"
  assert_false: expected false got true
[FAIL] MLActivation interface: existence and properties of interface object
  assert_own_property: self does not have own property "MLActivation" expected property "MLActivation" missing
[FAIL] MLActivation interface object length
  assert_own_property: self does not have own property "MLActivation" expected property "MLActivation" missing
[FAIL] MLActivation interface object name
  assert_own_property: self does not have own property "MLActivation" expected property "MLActivation" missing
[FAIL] MLActivation interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "MLActivation" expected property "MLActivation" missing
[FAIL] MLActivation interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "MLActivation" expected property "MLActivation" missing
[FAIL] MLActivation interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "MLActivation" expected property "MLActivation" missing
[FAIL] MLActivation must be primary interface of relu
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: relu is not defined"
[FAIL] Stringification of relu
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: relu is not defined"
[FAIL] MLContext interface: operation compute(MLGraph, MLNamedArrayBufferViews, MLNamedArrayBufferViews)
  assert_own_property: interface prototype object missing non-static operation expected property "compute" missing
[FAIL] MLContext interface: context must inherit property "compute(MLGraph, MLNamedArrayBufferViews, MLNamedArrayBufferViews)" with the proper type
  assert_inherits: property "compute" not found in prototype chain
[FAIL] MLContext interface: calling compute(MLGraph, MLNamedArrayBufferViews, MLNamedArrayBufferViews) on context with too few arguments must throw TypeError
  assert_inherits: property "compute" not found in prototype chain
[FAIL] MLGraph interface: existence and properties of interface object
  assert_own_property: self does not have own property "MLGraph" expected property "MLGraph" missing
[FAIL] MLGraph interface object length
  assert_own_property: self does not have own property "MLGraph" expected property "MLGraph" missing
[FAIL] MLGraph interface object name
  assert_own_property: self does not have own property "MLGraph" expected property "MLGraph" missing
[FAIL] MLGraph interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "MLGraph" expected property "MLGraph" missing
[FAIL] MLGraph interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "MLGraph" expected property "MLGraph" missing
[FAIL] MLGraph interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "MLGraph" expected property "MLGraph" missing
[FAIL] MLGraph must be primary interface of graph
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: graph is not defined"
[FAIL] Stringification of graph
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: graph is not defined"
[FAIL] MLGraphBuilder interface: existence and properties of interface object
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface object length
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface object name
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation input(DOMString, MLOperandDescriptor)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation constant(MLOperandDescriptor, ArrayBufferView)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation constant(double, optional MLOperandDataType)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation build(MLNamedOperands)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation argMin(MLOperand, optional MLArgMinMaxOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation argMax(MLOperand, optional MLArgMinMaxOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation batchNormalization(MLOperand, MLOperand, MLOperand, optional MLBatchNormalizationOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation cast(MLOperand, MLOperandDataType)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation clamp(MLOperand, optional MLClampOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation clamp(optional MLClampOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation concat(sequence<MLOperand>, unsigned long)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation conv2d(MLOperand, MLOperand, optional MLConv2dOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation convTranspose2d(MLOperand, MLOperand, optional MLConvTranspose2dOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation add(MLOperand, MLOperand)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation sub(MLOperand, MLOperand)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation mul(MLOperand, MLOperand)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation div(MLOperand, MLOperand)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation max(MLOperand, MLOperand)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation min(MLOperand, MLOperand)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation pow(MLOperand, MLOperand)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation equal(MLOperand, MLOperand)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation greater(MLOperand, MLOperand)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation greaterOrEqual(MLOperand, MLOperand)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation lesser(MLOperand, MLOperand)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation lesserOrEqual(MLOperand, MLOperand)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation not(MLOperand)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation abs(MLOperand)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation ceil(MLOperand)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation cos(MLOperand)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation erf(MLOperand)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation exp(MLOperand)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation floor(MLOperand)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation identity(MLOperand)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation log(MLOperand)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation neg(MLOperand)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation reciprocal(MLOperand)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation sin(MLOperand)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation sqrt(MLOperand)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation tan(MLOperand)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation elu(MLOperand, optional MLEluOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation elu(optional MLEluOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation expand(MLOperand, sequence<unsigned long>)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation gather(MLOperand, MLOperand, optional MLGatherOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation gemm(MLOperand, MLOperand, optional MLGemmOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation gru(MLOperand, MLOperand, MLOperand, unsigned long, unsigned long, optional MLGruOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation gruCell(MLOperand, MLOperand, MLOperand, MLOperand, unsigned long, optional MLGruCellOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation hardSigmoid(MLOperand, optional MLHardSigmoidOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation hardSigmoid(optional MLHardSigmoidOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation hardSwish(MLOperand)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation hardSwish()
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation instanceNormalization(MLOperand, optional MLInstanceNormalizationOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation layerNormalization(MLOperand, optional MLLayerNormalizationOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation leakyRelu(MLOperand, optional MLLeakyReluOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation leakyRelu(optional MLLeakyReluOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation linear(MLOperand, optional MLLinearOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation linear(optional MLLinearOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation lstm(MLOperand, MLOperand, MLOperand, unsigned long, unsigned long, optional MLLstmOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation lstmCell(MLOperand, MLOperand, MLOperand, MLOperand, MLOperand, unsigned long, optional MLLstmCellOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation matmul(MLOperand, MLOperand)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation pad(MLOperand, sequence<unsigned long>, sequence<unsigned long>, optional MLPadOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation averagePool2d(MLOperand, optional MLPool2dOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation l2Pool2d(MLOperand, optional MLPool2dOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation maxPool2d(MLOperand, optional MLPool2dOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation prelu(MLOperand, MLOperand)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation reduceL1(MLOperand, optional MLReduceOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation reduceL2(MLOperand, optional MLReduceOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation reduceLogSum(MLOperand, optional MLReduceOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation reduceLogSumExp(MLOperand, optional MLReduceOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation reduceMax(MLOperand, optional MLReduceOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation reduceMean(MLOperand, optional MLReduceOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation reduceMin(MLOperand, optional MLReduceOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation reduceProduct(MLOperand, optional MLReduceOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation reduceSum(MLOperand, optional MLReduceOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation reduceSumSquare(MLOperand, optional MLReduceOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation relu(MLOperand)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation relu()
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation resample2d(MLOperand, optional MLResample2dOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation reshape(MLOperand, sequence<unsigned long>)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation sigmoid(MLOperand)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation sigmoid()
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation slice(MLOperand, sequence<unsigned long>, sequence<unsigned long>)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation softmax(MLOperand)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation softmax()
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation softplus(MLOperand, optional MLSoftplusOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation softplus(optional MLSoftplusOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation softsign(MLOperand)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation softsign()
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation split(MLOperand, (unsigned long or sequence<unsigned long>), optional MLSplitOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation tanh(MLOperand)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation tanh()
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation transpose(MLOperand, optional MLTransposeOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation triangular(MLOperand, optional MLTriangularOptions)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder interface: operation where(MLOperand, MLOperand, MLOperand)
  assert_own_property: self does not have own property "MLGraphBuilder" expected property "MLGraphBuilder" missing
[FAIL] MLGraphBuilder must be primary interface of builder
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] Stringification of builder
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "input(DOMString, MLOperandDescriptor)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling input(DOMString, MLOperandDescriptor) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "constant(MLOperandDescriptor, ArrayBufferView)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling constant(MLOperandDescriptor, ArrayBufferView) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "constant(double, optional MLOperandDataType)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling constant(double, optional MLOperandDataType) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "build(MLNamedOperands)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling build(MLNamedOperands) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "argMin(MLOperand, optional MLArgMinMaxOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling argMin(MLOperand, optional MLArgMinMaxOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "argMax(MLOperand, optional MLArgMinMaxOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling argMax(MLOperand, optional MLArgMinMaxOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "batchNormalization(MLOperand, MLOperand, MLOperand, optional MLBatchNormalizationOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling batchNormalization(MLOperand, MLOperand, MLOperand, optional MLBatchNormalizationOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "cast(MLOperand, MLOperandDataType)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling cast(MLOperand, MLOperandDataType) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "clamp(MLOperand, optional MLClampOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling clamp(MLOperand, optional MLClampOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "clamp(optional MLClampOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling clamp(optional MLClampOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "concat(sequence<MLOperand>, unsigned long)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling concat(sequence<MLOperand>, unsigned long) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "conv2d(MLOperand, MLOperand, optional MLConv2dOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling conv2d(MLOperand, MLOperand, optional MLConv2dOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "convTranspose2d(MLOperand, MLOperand, optional MLConvTranspose2dOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling convTranspose2d(MLOperand, MLOperand, optional MLConvTranspose2dOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "add(MLOperand, MLOperand)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling add(MLOperand, MLOperand) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "sub(MLOperand, MLOperand)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling sub(MLOperand, MLOperand) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "mul(MLOperand, MLOperand)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling mul(MLOperand, MLOperand) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "div(MLOperand, MLOperand)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling div(MLOperand, MLOperand) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "max(MLOperand, MLOperand)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling max(MLOperand, MLOperand) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "min(MLOperand, MLOperand)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling min(MLOperand, MLOperand) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "pow(MLOperand, MLOperand)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling pow(MLOperand, MLOperand) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "equal(MLOperand, MLOperand)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling equal(MLOperand, MLOperand) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "greater(MLOperand, MLOperand)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling greater(MLOperand, MLOperand) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "greaterOrEqual(MLOperand, MLOperand)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling greaterOrEqual(MLOperand, MLOperand) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "lesser(MLOperand, MLOperand)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling lesser(MLOperand, MLOperand) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "lesserOrEqual(MLOperand, MLOperand)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling lesserOrEqual(MLOperand, MLOperand) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "not(MLOperand)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling not(MLOperand) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "abs(MLOperand)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling abs(MLOperand) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "ceil(MLOperand)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling ceil(MLOperand) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "cos(MLOperand)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling cos(MLOperand) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "erf(MLOperand)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling erf(MLOperand) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "exp(MLOperand)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling exp(MLOperand) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "floor(MLOperand)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling floor(MLOperand) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "identity(MLOperand)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling identity(MLOperand) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "log(MLOperand)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling log(MLOperand) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "neg(MLOperand)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling neg(MLOperand) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "reciprocal(MLOperand)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling reciprocal(MLOperand) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "sin(MLOperand)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling sin(MLOperand) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "sqrt(MLOperand)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling sqrt(MLOperand) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "tan(MLOperand)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling tan(MLOperand) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "elu(MLOperand, optional MLEluOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling elu(MLOperand, optional MLEluOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "elu(optional MLEluOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling elu(optional MLEluOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "expand(MLOperand, sequence<unsigned long>)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling expand(MLOperand, sequence<unsigned long>) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "gather(MLOperand, MLOperand, optional MLGatherOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling gather(MLOperand, MLOperand, optional MLGatherOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "gemm(MLOperand, MLOperand, optional MLGemmOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling gemm(MLOperand, MLOperand, optional MLGemmOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "gru(MLOperand, MLOperand, MLOperand, unsigned long, unsigned long, optional MLGruOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling gru(MLOperand, MLOperand, MLOperand, unsigned long, unsigned long, optional MLGruOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "gruCell(MLOperand, MLOperand, MLOperand, MLOperand, unsigned long, optional MLGruCellOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling gruCell(MLOperand, MLOperand, MLOperand, MLOperand, unsigned long, optional MLGruCellOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "hardSigmoid(MLOperand, optional MLHardSigmoidOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling hardSigmoid(MLOperand, optional MLHardSigmoidOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "hardSigmoid(optional MLHardSigmoidOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling hardSigmoid(optional MLHardSigmoidOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "hardSwish(MLOperand)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling hardSwish(MLOperand) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "hardSwish()" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "instanceNormalization(MLOperand, optional MLInstanceNormalizationOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling instanceNormalization(MLOperand, optional MLInstanceNormalizationOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "layerNormalization(MLOperand, optional MLLayerNormalizationOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling layerNormalization(MLOperand, optional MLLayerNormalizationOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "leakyRelu(MLOperand, optional MLLeakyReluOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling leakyRelu(MLOperand, optional MLLeakyReluOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "leakyRelu(optional MLLeakyReluOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling leakyRelu(optional MLLeakyReluOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "linear(MLOperand, optional MLLinearOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling linear(MLOperand, optional MLLinearOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "linear(optional MLLinearOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling linear(optional MLLinearOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "lstm(MLOperand, MLOperand, MLOperand, unsigned long, unsigned long, optional MLLstmOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling lstm(MLOperand, MLOperand, MLOperand, unsigned long, unsigned long, optional MLLstmOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "lstmCell(MLOperand, MLOperand, MLOperand, MLOperand, MLOperand, unsigned long, optional MLLstmCellOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling lstmCell(MLOperand, MLOperand, MLOperand, MLOperand, MLOperand, unsigned long, optional MLLstmCellOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "matmul(MLOperand, MLOperand)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling matmul(MLOperand, MLOperand) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "pad(MLOperand, sequence<unsigned long>, sequence<unsigned long>, optional MLPadOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling pad(MLOperand, sequence<unsigned long>, sequence<unsigned long>, optional MLPadOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "averagePool2d(MLOperand, optional MLPool2dOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling averagePool2d(MLOperand, optional MLPool2dOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "l2Pool2d(MLOperand, optional MLPool2dOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling l2Pool2d(MLOperand, optional MLPool2dOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "maxPool2d(MLOperand, optional MLPool2dOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling maxPool2d(MLOperand, optional MLPool2dOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "prelu(MLOperand, MLOperand)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling prelu(MLOperand, MLOperand) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "reduceL1(MLOperand, optional MLReduceOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling reduceL1(MLOperand, optional MLReduceOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "reduceL2(MLOperand, optional MLReduceOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling reduceL2(MLOperand, optional MLReduceOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "reduceLogSum(MLOperand, optional MLReduceOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling reduceLogSum(MLOperand, optional MLReduceOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "reduceLogSumExp(MLOperand, optional MLReduceOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling reduceLogSumExp(MLOperand, optional MLReduceOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "reduceMax(MLOperand, optional MLReduceOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling reduceMax(MLOperand, optional MLReduceOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "reduceMean(MLOperand, optional MLReduceOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling reduceMean(MLOperand, optional MLReduceOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "reduceMin(MLOperand, optional MLReduceOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling reduceMin(MLOperand, optional MLReduceOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "reduceProduct(MLOperand, optional MLReduceOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling reduceProduct(MLOperand, optional MLReduceOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "reduceSum(MLOperand, optional MLReduceOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling reduceSum(MLOperand, optional MLReduceOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "reduceSumSquare(MLOperand, optional MLReduceOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling reduceSumSquare(MLOperand, optional MLReduceOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "relu(MLOperand)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling relu(MLOperand) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "relu()" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "resample2d(MLOperand, optional MLResample2dOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling resample2d(MLOperand, optional MLResample2dOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "reshape(MLOperand, sequence<unsigned long>)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling reshape(MLOperand, sequence<unsigned long>) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "sigmoid(MLOperand)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling sigmoid(MLOperand) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "sigmoid()" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "slice(MLOperand, sequence<unsigned long>, sequence<unsigned long>)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling slice(MLOperand, sequence<unsigned long>, sequence<unsigned long>) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "softmax(MLOperand)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling softmax(MLOperand) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "softmax()" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "softplus(MLOperand, optional MLSoftplusOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling softplus(MLOperand, optional MLSoftplusOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "softplus(optional MLSoftplusOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling softplus(optional MLSoftplusOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "softsign(MLOperand)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling softsign(MLOperand) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "softsign()" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "split(MLOperand, (unsigned long or sequence<unsigned long>), optional MLSplitOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling split(MLOperand, (unsigned long or sequence<unsigned long>), optional MLSplitOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "tanh(MLOperand)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling tanh(MLOperand) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "tanh()" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "transpose(MLOperand, optional MLTransposeOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling transpose(MLOperand, optional MLTransposeOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "triangular(MLOperand, optional MLTriangularOptions)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling triangular(MLOperand, optional MLTriangularOptions) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: builder must inherit property "where(MLOperand, MLOperand, MLOperand)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLGraphBuilder interface: calling where(MLOperand, MLOperand, MLOperand) on builder with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: builder is not defined"
[FAIL] MLOperand interface: existence and properties of interface object
  assert_own_property: self does not have own property "MLOperand" expected property "MLOperand" missing
[FAIL] MLOperand interface object length
  assert_own_property: self does not have own property "MLOperand" expected property "MLOperand" missing
[FAIL] MLOperand interface object name
  assert_own_property: self does not have own property "MLOperand" expected property "MLOperand" missing
[FAIL] MLOperand interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "MLOperand" expected property "MLOperand" missing
[FAIL] MLOperand interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "MLOperand" expected property "MLOperand" missing
[FAIL] MLOperand interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "MLOperand" expected property "MLOperand" missing
[FAIL] MLOperand interface: operation dataType()
  assert_own_property: self does not have own property "MLOperand" expected property "MLOperand" missing
[FAIL] MLOperand interface: operation shape()
  assert_own_property: self does not have own property "MLOperand" expected property "MLOperand" missing
[FAIL] MLOperand must be primary interface of input
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: input is not defined"
[FAIL] Stringification of input
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: input is not defined"
[FAIL] MLOperand interface: input must inherit property "dataType()" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: input is not defined"
[FAIL] MLOperand interface: input must inherit property "shape()" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: input is not defined"
[FAIL] MLOperand must be primary interface of filter
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: filter is not defined"
[FAIL] Stringification of filter
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: filter is not defined"
[FAIL] MLOperand interface: filter must inherit property "dataType()" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: filter is not defined"
[FAIL] MLOperand interface: filter must inherit property "shape()" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: filter is not defined"
[FAIL] MLOperand must be primary interface of output
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: output is not defined"
[FAIL] Stringification of output
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: output is not defined"
[FAIL] MLOperand interface: output must inherit property "dataType()" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: output is not defined"
[FAIL] MLOperand interface: output must inherit property "shape()" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: output is not defined"
Harness: the test ran to completion.
